Fix/setup (#2)
* fix(devcontainer): setup

* fix(typing): some issues in class var types

* fix(devcontainer): ruff configurations

* fix(devcontainer): gitconfig

+++ justfile
@@ -4,8 +4,11 @@ default:
 init:
     #!/bin/zsh
     set -euxo pipefail
+    if [ -f /tmp/.gitconfig ]; then rm -rf /home/vscode/.gitconfig && cp /tmp/.gitconfig /home/vscode/.gitconfig; fi
     git config --global --add safe.directory '*'
-    cp /tmp/.zshrc /home/vscode/.zshrc
+    git config --global core.autocrlf input
+    git config --global core.eol lf
+    if [ -f /tmp/.zshrc ]; then cp /tmp/.zshrc /home/vscode/.zshrc; fi
     rm -rf .venv
     uv venv
     uv sync --all-groups
